pyproject.toml
[project]
name = "breaking-ai-systems"
version = "0.1.0"
description = "Breaking AI Systems: From Image Classifiers to LLM Agents"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
authors = [
    { name = "Raja Sekhar Rao Dheekonda" },
]
keywords = [
    "adversarial-attacks",
    "ai-safety",
    "ai-security",
    "jailbreaking",
    "llm",
    "red-teaming",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
]
dependencies = [
    "requests>=2.31",
    "pillow>=10.0",
    "numpy>=1.26",
    "matplotlib>=3.8",
    "pandas>=2.1",
    "ipykernel>=6.29",
    "ipywidgets>=8.1",
    "torchvision>=0.17",
    "adversarial-robustness-toolbox>=1.18",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "litellm>=1.40",
]

[project.urls]
Repository = "https://github.com/rdheekonda/breaking-ai-systems"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["core"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
files = ["core"]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
